+++ CHANGELOG.md
@@ -3,7 +3,7 @@
   This may be explicitly overridden by passing `--overwrite-input`.
 * Add crf-search,auto-encode arg `--high-crf-means-hq` for use with encoders where high 
   crf means _higher_ quality.
-* Default `--high-crf-means-hq` for hevc_videotoolbox.
+* Encoder hevc_videotoolbox: Default `--high-crf-means-hq`, map `--crf` to `-q:v`.
 
 # v0.10.1
 * Support setting `--enc-input hwaccel=none --enc-input hwaccel_output_format=none` to omit defaults